--- strelyaev.roman/P2/checkargs.h
@@ -1,6 +1,0 @@
-#ifndef CHECKARGS_H
-#define CHECKARGS_H
-#include <fstream>
-void checkArgs(int argc, char * argv[], int n, std::fstream & in);
-#endif
-
